c3_rnt2_ai/pyproject.toml
[project]
name = "c3-rnt2-ai"
version = "0.1.0"
description = "Local C3 + RNT-2 + KV-hybrid research stack"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Local", email = "local@example.com"}]
license = {text = "MIT"}
dependencies = [
  "torch>=2.1",
  "typer>=0.9",
  "rich>=13",
  "pyyaml>=6",
  "numpy>=1.24",
  "requests>=2.31",
  "zstandard>=0.22",
  "lz4>=4.3",
  "faiss-cpu>=1.7",
  "tqdm>=4.66"
]

[project.optional-dependencies]
api = ["fastapi>=0.110", "uvicorn>=0.23"]
triton = ["triton>=2.1"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
